@@ -1,39 +1,88 @@
+import { CssBaseline, ThemeProvider, StyledEngineProvider } from "@mui/material";
+import { FC, ReactElement } from "react";
 import { HashRouter, Route, Switch } from "react-router-dom";
-import React, { FC, ReactElement, useState } from "react";
-import styled from "styled-components";
-import Home from "./pages/Home";
-import { CssBaseline, ThemeProvider } from "@material-ui/core";
-import 'typeface-roboto';
-import MainWizard from "./pages/MainWizard";
-import theme from "./theme";
-import { Network } from './types';
 
-const Container = styled.main`
-  display: flex;
-  flex-direction: column;
-  min-height: 100vh;
-`;
+import BTECContextWrapper from "./BTECContext";
+import { OnlineDetector } from "./components/OnlineDetector";
+import VersionFooter from "./components/VersionFooter";
+import { paths } from "./constants";
+import GlobalContextWrapper from "./GlobalContext";
+import KeyCreationContextWrapper from "./KeyCreationContext";
+import ConfigureValidatorKeys from "./pages/ConfigureValidatorKeys";
+import ConfigureWithdrawalAddress from "./pages/ConfigureWithdrawalAddress";
+import CreateCredentialsChange from "./pages/CreateCredentialsChange";
+import CreateMnemonic from "./pages/CreateMnemonic";
+import CreateValidatorKeys from "./pages/CreateValidatorKeys";
+import FinishCredentialsGeneration from "./pages/FinishCredentialsGeneration";
+import FinishKeyGeneration from "./pages/FinishKeyGeneration";
+import Home from "./pages/Home";
+import MnemonicImport from "./pages/MnemonicImport";
+import theme from "./theme";
 
 /**
- * The React app top level including theme and routing.
- * 
- * @returns the react element containing the app
+ * Routing for the application. Broken into four sections:
+ * - Primary home page
+ * - Routes for creating a mnemonic and validator keys
+ * - Routes for using an existing mnemonic to create validator keys
+ * - Routes for generating the credentials change
+ *
+ * Each of the three flows is wrapped in a React Context that will store
+ * the inputs of the user to be accessible across each page. This prevents
+ * prop drilling
  */
 const App: FC = (): ReactElement => {
-  const [network, setNetwork] = useState<Network>(Network.GNOSIS);
-
   return (
-    <ThemeProvider theme={theme}>
-      <CssBaseline />
-      <HashRouter>
-        <Container>
-          <Switch>
-            <Route exact path="/" render={() => <Home network={network} setNetwork={setNetwork} />} />
-            <Route exact path="/wizard/:stepSequenceKey" render={() => <MainWizard network={network} />} />
-          </Switch>
-        </Container>
-      </HashRouter>
-    </ThemeProvider>
+    <StyledEngineProvider injectFirst>
+      <ThemeProvider theme={theme}>
+        <CssBaseline />
+        <HashRouter>
+          <GlobalContextWrapper>
+            <main className="tw-flex tw-flex-col tw-h-[100vh]">
+              <OnlineDetector />
+              <Switch>
+                <Route exact path="/" render={() => <Home />} />
+
+                <Route>
+
+                  {/* Create Mnemonic & Keys Flow */}
+                  <KeyCreationContextWrapper>
+                    <Switch>
+                      <Route path={paths.CREATE_MNEMONIC} children={() => <CreateMnemonic />} />
+                      <Route path={paths.CONFIGURE_CREATE} children={() => <ConfigureValidatorKeys />} />
+                      <Route path={paths.CREATE_KEYS_CREATE} children={() => <CreateValidatorKeys />} />
+                      <Route path={paths.FINISH_CREATE} children={() => <FinishKeyGeneration />} />
+                    </Switch>
+                  </KeyCreationContextWrapper>
+
+
+                  {/* Import Mnemonic & Generate Keys Flow */}
+                  <KeyCreationContextWrapper>
+                    <Switch>
+                      <Route path={paths.EXISTING_IMPORT} render={() => <MnemonicImport />} />
+                      <Route path={paths.CONFIGURE_EXISTING} render={() => <ConfigureValidatorKeys />} />
+                      <Route path={paths.CREATE_KEYS_EXISTING} render={() => <CreateValidatorKeys />} />
+                      <Route path={paths.FINISH_EXISTING} render={() => <FinishKeyGeneration />} />
+                    </Switch>
+                  </KeyCreationContextWrapper>
+
+
+                  {/* Update Withdrawal Credentials Flow */}
+                  <BTECContextWrapper>
+                    <Switch>
+                      <Route path={paths.BTEC_IMPORT} render={() => <MnemonicImport />} />
+                      <Route path={paths.CONFIGURE_BTEC} render={() => <ConfigureWithdrawalAddress />} />
+                      <Route path={paths.CREATE_CREDENTIALS} render={() => <CreateCredentialsChange />} />
+                      <Route path={paths.FINISH_CREDENTIALS} render={() => <FinishCredentialsGeneration />} />
+                    </Switch>
+                  </BTECContextWrapper>
+                </Route>
+              </Switch>
+              <VersionFooter />
+            </main>
+          </GlobalContextWrapper>
+        </HashRouter>
+      </ThemeProvider>
+    </StyledEngineProvider>
   );
 };
 
