@@ -1,165 +1,152 @@
+import { Button, Tooltip, Typography } from "@mui/material";
+import { useContext, useEffect, useMemo, useState } from "react";
 import { useHistory } from "react-router-dom";
-import React, { FC, ReactElement, useState, Dispatch, SetStateAction } from "react";
-import styled from "styled-components";
-import { Container, Grid, Modal, Tooltip, Typography } from "@material-ui/core";
-import { Button } from '@material-ui/core';
-import { KeyIcon } from "../components/icons/KeyIcon";
-import { NetworkPicker } from "../components/NetworkPicker";
-import { tooltips } from "../constants";
-import { Network, StepSequenceKey } from '../types'
-import VersionFooter from "../components/VersionFooter";
 
-const StyledMuiContainer = styled(Container)`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-`;
-
-const NetworkDiv = styled.div`
-  margin-top: 35px;
-  margin-right: 35px;
-  align-self: flex-end;
-  color: gray;
-`;
-
-const LandingHeader = styled(Typography)`
-  font-size: 36px;
-  margin-top: 15px;
-  margin-bottom: 20px;
-`;
-
-const SubHeader = styled(Typography)`
-  margin-top: 20px;
-`;
-
-const Links = styled.div`
-  margin-top: 35px;
-`;
-
-const InfoLabel = styled.span`
-  color: gray;
-`;
-
-const OptionsGrid = styled(Grid)`
-  margin-top: 35px;
-  align-items: center;
-`;
-
-type HomeProps = {
-  network: Network,
-  setNetwork: Dispatch<SetStateAction<Network>>
-}
+import { paths, tooltips } from "../constants";
+import { GlobalContext } from "../GlobalContext";
+import { KeyIcon } from "../icons/KeyIcon";
+import NetworkPickerModal from "../modals/NetworkPickerModal";
+import ReuseMnemonicActionModal from "../modals/ReuseMnemonicActionModal";
+import { ReuseMnemonicAction } from "../types";
 
 /**
- * Home page and entry point of the app.  This page displays general information
- * and options for a user to create a new secret recovery phrase or use an 
- * existing one.
- * 
- * @param props passed in data for the component to use
- * @returns the react element to render
+ * Landed page of the application.
+ * The user will be able to select a network and choose the primary action
+ * they wish to make.
  */
-const Home: FC<HomeProps> = (props): ReactElement => {
+const Home = () => {
+  const { network } = useContext(GlobalContext);
+  const [wasNetworkModalOpened, setWasNetworkModalOpened] = useState(false);
   const [showNetworkModal, setShowNetworkModal] = useState(false);
-  const [networkModalWasOpened, setNetworkModalWasOpened] = useState(true);
+  const [showReuseMnemonicModal, setShowReuseMnemonicModal] = useState(false);
   const [createMnemonicSelected, setCreateMnemonicSelected] = useState(false);
   const [useExistingMnemonicSelected, setUseExistingMnemonicSelected] = useState(false);
 
   let history = useHistory();
 
+  const tabIndex = useMemo(() => showNetworkModal ? -1 : 1, [showNetworkModal]);
+
   const handleOpenNetworkModal = () => {
     setShowNetworkModal(true);
-    setNetworkModalWasOpened(true);
-  }
+    setWasNetworkModalOpened(true);
+  };
 
-  const handleCloseNetworkModal = (event: object, reason: string) => {
-    if (reason !== 'backdropClick') {
-      setShowNetworkModal(false);
-
-      if (createMnemonicSelected) {
-        handleCreateNewMnemonic();
-      } else if (useExistingMnemonicSelected) {
-        handleUseExistingMnemonic();
-      }
+  const handleCloseNetworkModal = () => {
+    setShowNetworkModal(false);
+    if (createMnemonicSelected) {
+      handleCreateNewMnemonic();
+    } else if (useExistingMnemonicSelected) {
+      handleUseExistingMnemonic();
     }
-  }
+  };
 
   const handleCreateNewMnemonic = () => {
     setCreateMnemonicSelected(true);
 
-    if (!networkModalWasOpened) {
+    if (!wasNetworkModalOpened) {
       handleOpenNetworkModal();
     } else {
-      const location = {
-        pathname: `/wizard/${StepSequenceKey.MnemonicGeneration}`
-      }
-
-      history.push(location);
+      history.push(paths.CREATE_MNEMONIC)
     }
-  }
+  };
 
   const handleUseExistingMnemonic = () => {
     setUseExistingMnemonicSelected(true);
 
-    if (!networkModalWasOpened) {
+    if (!wasNetworkModalOpened) {
       handleOpenNetworkModal();
     } else {
-      const location = {
-        pathname: `/wizard/${StepSequenceKey.MnemonicImport}`
-      }
-
-      history.push(location);
+      setShowReuseMnemonicModal(true);
     }
-  }
+  };
 
-  const tabIndex = (priority: number) => showNetworkModal ? -1 : priority;
+  const handleCloseReuseActionModal = () => {
+    setShowReuseMnemonicModal(false);
+  };
+
+  const handleReuseMnemonicActionSubmit = (action: ReuseMnemonicAction) => {
+    setShowReuseMnemonicModal(false);
+    if (action === ReuseMnemonicAction.RegenerateKeys) {
+
+      history.push(paths.EXISTING_IMPORT);
+    } else if (action === ReuseMnemonicAction.GenerateBLSToExecutionChange) {
+
+      history.push(paths.BTEC_IMPORT);
+    }
+  };
 
   return (
-    <StyledMuiContainer>
-      <NetworkDiv>
-        Select Network: <Button color="primary" onClick={handleOpenNetworkModal} tabIndex={tabIndex(1)}>{props.network}</Button>
-      </NetworkDiv>
-      <Modal
-        open={showNetworkModal}
-        onClose={handleCloseNetworkModal}
-      >
-        {/* Added <div> here per the following link to fix error https://stackoverflow.com/a/63521049/5949270 */}
-        <div>
-          <NetworkPicker handleCloseNetworkModal={handleCloseNetworkModal} setNetwork={props.setNetwork} network={props.network}></NetworkPicker>
+    <div className="tw-flex tw-flex-col tw-pt-8">
+      <div className="tw-pr-8 tw-text-right tw-w-full">
+        <span className="tw-text-gray tw-text-sm">Select Network:</span>{" "}
+        <Button
+          color="primary"
+          onClick={handleOpenNetworkModal}
+          tabIndex={tabIndex}
+        >
+          {network}
+        </Button>
+      </div>
+
+      <div className="tw-flex tw-flex-col tw-items-center tw-mt-4">
+        <Typography className="tw-text-4xl tw-mb-5" variant="h1">Welcome!</Typography>
+
+        <KeyIcon />
+
+        <Typography className="tw-mt-2">Your key generator for staking on Ethereum.</Typography>
+        <Typography>
+          You should run this tool{" "}
+          <Tooltip title={tooltips.OFFLINE}>
+            <span className="tw-underline">offline</span>
+          </Tooltip>{" "}
+          for your own security.
+        </Typography>
+
+        <div className="tw-mt-5">
+          <div>
+            <span className="tw-text-gray">Github:</span>{" "}
+            <span className="tw-text-sm">https://github.com/stake-house/wagyu-key-gen</span>
+          </div>
+          <div>
+            <span className="tw-text-gray">Support:</span>{" "}
+            <span className="tw-text-sm">https://dsc.gg/ethstaker</span>
+          </div>
         </div>
-      </Modal>
 
-      <LandingHeader variant="h1"><b>Welcome to the Gnosis Wagyu Key Generator</b></LandingHeader> 
-      <KeyIcon />
-      <SubHeader variant="h2"><b>Your Key Generator for The Gnosis Beacon Chain</b></SubHeader>
-      <SubHeader variant="h2"><i>This Program is in Final Beta Testing; Please Use At Your Own Risk!</i></SubHeader>
+        <Button
+          variant="contained"
+          color="primary"
+          className="tw-mt-5"
+          onClick={handleCreateNewMnemonic}
+          tabIndex={tabIndex}
+        >
+          Create New Secret Recovery Phrase
+        </Button>
 
-      <Links>
-        <InfoLabel>Github:</InfoLabel> https://github.com/alexpeterson91/wagyu-key-gen/issues
-        <br />
-        <InfoLabel>Support:</InfoLabel> https://discord.gg/dappnode
-        <br />
-        <InfoLabel>Powered by Stake-House:</InfoLabel> https://github.com/stake-house/wagyu-key-gen
-
-      </Links>
-
-      <OptionsGrid container spacing={2} direction="column">
-        <Grid item>
-          <Button variant="contained" color="primary" onClick={handleCreateNewMnemonic} tabIndex={tabIndex(1)}>
-            Create New Mnemonic Recovery Phrase
+        <Tooltip title={tooltips.IMPORT_MNEMONIC}>
+          <Button
+            className="tw-text-gray tw-mt-2"
+            size="small"
+            onClick={handleUseExistingMnemonic}
+            tabIndex={tabIndex}
+          >
+            Use Existing Secret Recovery Phrase
           </Button>
-        </Grid>
-        <Grid item>
-          <Tooltip title={tooltips.IMPORT_MNEMONIC}>
-            <Button variant="contained" color="primary" onClick={handleUseExistingMnemonic} tabIndex={tabIndex(1)}>
-              Use Existing Mnemonic Recovery Phrase
-            </Button>
-          </Tooltip>
-        </Grid>
-      </OptionsGrid>
-      <VersionFooter />
-    </StyledMuiContainer>
-  );
+        </Tooltip>
+      </div>
+
+      <NetworkPickerModal
+        onClose={handleCloseNetworkModal}
+        showModal={showNetworkModal}
+      />
+
+      <ReuseMnemonicActionModal
+        onClose={handleCloseReuseActionModal}
+        onSubmit={handleReuseMnemonicActionSubmit}
+        showModal={showReuseMnemonicModal}
+      />
+    </div>
+  )
 };
 
 export default Home;
