@@ -3,14 +3,24 @@ export enum StepKey {
   MnemonicGeneration,
   KeyConfiguration,
   KeyGeneration,
-  Finish
+  Finish,
+  BTECConfiguration,
+  BTECGeneration,
+  FinishBTEC
 }
 
 export enum StepSequenceKey {
   MnemonicGeneration = "mnemonicgeneration",
-  MnemonicImport = "mnemonicimport"
+  MnemonicImport = "mnemonicimport",
+  BLSToExecutionChangeGeneration = "blstoexecutionchangegeneration",
+}
+
+export enum ReuseMnemonicAction {
+  RegenerateKeys,
+  GenerateBLSToExecutionChange
 }
 
 export enum Network {
-  GNOSIS = "Gnosis"
+  GNOSIS = "Gnosis",
+  CHIADO = "Chiado"
 }
